Refactor `configuration` package
commit-id:788f97c7
