+++ crates/sncast/src/starknet_commands/account/mod.rs
@@ -335,7 +335,7 @@ pub async fn account(
 #[cfg(test)]
 mod tests {
     use camino::Utf8PathBuf;
-    use configuration::copy_config_to_tempdir;
+    use configuration::test_utils::copy_config_to_tempdir;
     use sncast::helpers::configuration::CastConfig;
     use sncast::helpers::constants::DEFAULT_ACCOUNTS_FILE;
     use std::fs;
@@ -344,8 +344,7 @@ mod tests {
 
     #[test]
     fn test_add_created_profile_to_configuration_happy_case() {
-        let tempdir =
-            copy_config_to_tempdir("tests/data/files/correct_snfoundry.toml", None).unwrap();
+        let tempdir = copy_config_to_tempdir("tests/data/files/correct_snfoundry.toml", None);
         let path = Utf8PathBuf::try_from(tempdir.path().to_path_buf()).unwrap();
         let config = CastConfig {
             url: String::from("http://some-url"),
@@ -370,8 +369,7 @@ mod tests {
 
     #[test]
     fn test_add_created_profile_to_configuration_profile_already_exists() {
-        let tempdir =
-            copy_config_to_tempdir("tests/data/files/correct_snfoundry.toml", None).unwrap();
+        let tempdir = copy_config_to_tempdir("tests/data/files/correct_snfoundry.toml", None);
         let config = CastConfig {
             url: String::from("http://127.0.0.1:5055/rpc"),
             account: String::from("user1"),
