+++ crates/sncast/tests/e2e/main_tests.rs
@@ -7,13 +7,13 @@ use crate::helpers::fixtures::{
     duplicate_contract_directory_with_salt, get_accounts_path, get_keystores_path,
 };
 use crate::helpers::runner::runner;
-use configuration::copy_config_to_tempdir;
+use configuration::test_utils::copy_config_to_tempdir;
 use indoc::indoc;
 use shared::test_utils::output_assert::assert_stderr_contains;
 
 #[tokio::test]
 async fn test_happy_case_from_sncast_config() {
-    let tempdir = copy_config_to_tempdir("tests/data/files/correct_snfoundry.toml", None).unwrap();
+    let tempdir = copy_config_to_tempdir("tests/data/files/correct_snfoundry.toml", None);
     let args = vec![
         "--accounts-file",
         ACCOUNT_FILE_PATH,
@@ -37,7 +37,7 @@ async fn test_happy_case_from_sncast_config() {
 
 #[tokio::test]
 async fn test_happy_case_predefined_network() {
-    let tempdir = copy_config_to_tempdir("tests/data/files/correct_snfoundry.toml", None).unwrap();
+    let tempdir = copy_config_to_tempdir("tests/data/files/correct_snfoundry.toml", None);
     let args = vec![
         "--accounts-file",
         ACCOUNT_FILE_PATH,
@@ -63,7 +63,7 @@ async fn test_happy_case_predefined_network() {
 
 #[tokio::test]
 async fn test_url_with_network_args() {
-    let tempdir = copy_config_to_tempdir("tests/data/files/correct_snfoundry.toml", None).unwrap();
+    let tempdir = copy_config_to_tempdir("tests/data/files/correct_snfoundry.toml", None);
     let args = vec![
         "--accounts-file",
         ACCOUNT_FILE_PATH,
@@ -91,7 +91,7 @@ async fn test_url_with_network_args() {
 
 #[tokio::test]
 async fn test_network_with_url_defined_in_config_toml() {
-    let tempdir = copy_config_to_tempdir("tests/data/files/correct_snfoundry.toml", None).unwrap();
+    let tempdir = copy_config_to_tempdir("tests/data/files/correct_snfoundry.toml", None);
     let args = vec![
         "--accounts-file",
         ACCOUNT_FILE_PATH,
@@ -142,7 +142,7 @@ async fn test_happy_case_from_cli_no_scarb() {
 
 #[tokio::test]
 async fn test_happy_case_from_cli_with_sncast_config() {
-    let tempdir = copy_config_to_tempdir("tests/data/files/correct_snfoundry.toml", None).unwrap();
+    let tempdir = copy_config_to_tempdir("tests/data/files/correct_snfoundry.toml", None);
     let args = vec![
         "--accounts-file",
         ACCOUNT_FILE_PATH,
@@ -175,7 +175,7 @@ async fn test_happy_case_from_cli_with_sncast_config() {
 
 #[tokio::test]
 async fn test_happy_case_mixed() {
-    let tempdir = copy_config_to_tempdir("tests/data/files/correct_snfoundry.toml", None).unwrap();
+    let tempdir = copy_config_to_tempdir("tests/data/files/correct_snfoundry.toml", None);
     let args = vec![
         "--accounts-file",
         ACCOUNT_FILE_PATH,
